Made dictionary local to method and move model to domain folder
